Merge in ASDF/ese_parser from ASDF-5116 to main

* commit '43a62d9a34d215d3c02c87fbf5a710cc7559ddff':
  B!! Fix u32 overflow issue

+++ lib/src/parser/jet.rs
@@ -311,7 +311,7 @@ impl DbPage {
     }
 
     pub fn offset(&self) -> u64 {
-        ((self.page_number + 1) * self.page_size) as u64
+        (self.page_number as u64 + 1) * self.page_size as u64
     }
 }
